feat(api): Laying groundwork for camt.053 support
This adds the schemas needed to parse v2 and v8 files as well as sample
files obtained from goldman sachs. This also adds some tests to make
sure the files can be parsed at a most basic level. Next steps will be
getting useful information out of the files in a way that monetr can use
them.

Related to #2352