fix(api): Better scrub sentry data
Sentry in prod is already configured to scrub sensitive data and
identifying things. But this might not be true for people self hosting.
So for that use case this will make it so that you don't really need to
configure sentry to scrub data. Sensitive data simply will not be
sent to sentry.