feat(api): Improving deleting transactions
This moves the transaction delete options into two query parameters.
`adjusts_balance` which if a truthy value is provided will modify the
balances when the transaction is removed (including putting funds back
into the spending object if the transaction has one). As well as adding
a `soft` query parameter. This defaults to a truthy value, but if a
falsey value is specified then the transaction will be hard deleted from
the database.

Related to #2542