feat(ui): Adding UX to remove transactions for manual links (#2564)
* feat(ui): Adding UX to remove transactions for manual links

If you are working with a manual budget, you will now be able to remove
transactions from that budget. Navigating to the transaction details
page and clicking remove will show a confirmation modal, prompting you
whether or not removing the transactionw will affect balances.

* chore: REmoving more mui icons

* chore: Ignore pnpm scripts

* feat(ui): Adding remove transaction modal

Added a modal to confirm removing a transaction as well as setting the
optional params for the removal.

Improved some styling around the top navigation and other items.

* fix(ui): Improve UI around remove txn button