fix(api): Fix importing an OFX file missing curdef
If an OFX file is missing the curdef field we should fallback to the
bank account's currency that we are importing for. This will resolve an
error when the currency is on a per transaction basis rather than for
the entire file.

This also improves the parsing of OFX dates allowing for even more
formats and the ability to easily add more formats.

Resolves #2575