fix(plaid): Fix rounding error in Plaid amounts
Because monetr does not store floating point numbers (because they are
the work of the devil), it instead converts floating point numbers to
whole integers. But this process is also sinful and can be error prone.
The way monetr was converting the amounts from Plaid was not consistent
with how the amounts were being converted elsewhere, and thus introduced
a problem where the amount monetr would indicate could be 0.01 off from
the amount Plaid or the real bank account indicated. This corrects the
issue by making sure that the Plaid code also uses the same code path as
other amount parsing code.

Resolves #2594