fix(ui): Use account timezone if available instead of browser timezone (#2390)
* fix(ui): Use account timezone if available instead of browser timezone

If a user has their timezone configured as X and their browser timezone
is Y. This can cause some of the dates sent to the server to be
malformed and can cause some odd behaviors. For lack of a better
solution for now (like not having times at all and instead only using
dates?); this adds timezone to components that need it. Forcing the time
to be shown and calculated in the user's timezone instead of only the
browsers timezone.

Resolves #2386

* chore: Fix issue with select

* chore: Progress on timezone things

This gets this PR a lot closer to being finished. Basically every place
where we send a date to the API backend we need to make sure that the
date has been adjusted to be in the users timezone regardless of the
timezone the user is currently in.