fix(api): Adding test for miscalculation in spending code
This adds a test a few minor tweaks to the spending calculation code.
I'm unable to reproduce the problem in the issue though so I'm not sure
exaclty what happened here. But I think overall it would be best to
continue trying to move away from this code path entirely and instead
move to the forecasting code path which is easier to maintain and test.

Resolves #2623