fix(ui): More timezone bug fixes (#2635)
Basically if you are in a different timezone than your actual account
then a lot of the UI displays incorrectly and the API can behave in odd
ways when you submit forms with dates in the wrong timezone.

This corrects even more of those issues and also removes timezone
specific dates entirely from the database as they weren't even being
used. Everything was always being stored in UTC and then was converted
to the user's timezone when the timestamp was actually being used. This
is far more effective anyway and is easier to test against as there are
fewer ways for data to get into the database wrong. Since all code paths
can be tested for correctness now.