fix(api): Fixing error logs and bad backoff on retry
Right now in production, we are getting error logs for context canceled
which is creating noise in our alerts. These errors are not necessarily
helpful and so I'm demoting them to be a warning instead.

I'm also adding a small backoff sleep to the Plaid sync retries so that
way we don't just keep retrying instantly. Though the fact that I'm
hitting the pagination error from Plaid so much on its own is pretty
fucking stupid. Like you sent me the webhook? How has it changed
already?