fix(build): Fixed vault in local development
For some reason the vault API request that monetr makes upon startup
would always fail if it was the first API call it made. But subsequent
API calls would succeed? I have no idea why this is however these
changes should help for some reason.

I've removed the data storage for vault, which means for local dev at
least; if vault is shutdown or restarted. All encryption is lost.
However this seems to fix the issue with volume mount permissioning on
linux systems and may have contributed to the error I was seeing.

Beyond that I also added a round tripper to vault which will
automatically instrument sentry traces for it as well.

Resolves #2787