chore(build): Removing flipt from development environment
I still want to build a feature flag solution into monetr because there
are things that I want to be able to test in prod or allow users to
test on their own as a "Beta" but without giving it to everyone and
setting the expectation that it just works.

But flipt made some changes in their v2 release, mostly moving away from
postgresql storage for their backend; which is a deal breaker for
monetr's production environment.