refactor(ui): Switched button component to CSS modules
This is the start of an effort to move away from tailwind, the upgrade
to tailwind v4 basically means that all of the configuration is in CSS
anyway. And the components here are already tricky to maintain in that
the tailwind styles are a mess. If there is going to be an effort to
consolidate and make all the components consistent. This seems like a
better use of time.