refactor(ui): Removing zustand, as it was only used for mobile sidebar
The state of the mobile sidebar (allowing it to be opened or closed by a
component that was nowhere near it in the tree) was managed via zustand.
This was overkill and I have updated this to instead just use react
context instead as thats really all this needs to be.

fix(ui): Fixed icons rendering uncolored

This was happening because emotion was finally removed.