feat(api): Improve similar transaction calculations
This significantly improves similar transaction calculations as well as
significantly improves merchant name recognition as a result of it. This
precedes changes where transaction clusters are not truncated before
every update. Instead transaction clusters will soon be persistent and
as a result they will be far more powerful tools that can be used for a
huge number of features and automations.