chore: Tinkering
This changes how the merchant name is calculated, instead the word
indicies are collected in the first iteration. Then the centermost point
of the cluster is calculated, from there the merchant name is caluclated
using the values of the centermost point in the cluster.

The centermost point is also deduplicated because if there are multiple
transactions with identical transaction names then they will all share
the same coordinates. This is deduplicated by preferring the transaction
with the lowest ID of the cluster for the center point.