fix(api): Fixing similar transaction panic
There is a bug in the similar transaction post-calculations where we
sort the vector by its value in order to get the highest value words in
a transaction. But it is possible for this vector to have negative
values which are meaningful but don't get sorted properly. So now the
sorting uses the absolute values of these items instead to avoid this
problem.

Resolves #2833