feat(api): Adding Lunch Flow as a datasource
Lunch Flow provides a way for users who are self hosting their own
budgeting apps to sync their data with institutions that are in the
EU/UK as well as other countries as well. This was previously very
difficult to do as an individual.

This is adding support for Lunch Flow as a datasource for monetr, but it
will only be available for self hosted instances.

Related to #2427