fix(ui): Handle authentication errors properly
If we get a 500 status code or something similar from the
`/api/users/me` endpoint then we want to handle this gracefully. We do
not want to accidently show something to the user where it may incidate
that they are not subscribed or something like that or some other odd
screen state.