fix(ui): Trying another fix for iOS and drawer issues
What is happening is basically on iOS when the view has been scrolled
past a certain amount the drawer showing up will cause the view to snap
to the top as if we've scrolled back to the top. But the transition is
jarring and even though the background is faded a bit the change is not
great and is noticable.

Trying to resolve this.