docs(blog): Blog post on monetr's similar transactions (#2509)
* docs(blog): Blog post on monetr's similar transactions

This is a working branch with some code I'll need for similar
transactions as well as the base blog post file.

* chore: Tweaks

* chore: Updating

* chore: Tinkering

This changes how the merchant name is calculated, instead the word
indicies are collected in the first iteration. Then the centermost point
of the cluster is calculated, from there the merchant name is caluclated
using the values of the centermost point in the cluster.

The centermost point is also deduplicated because if there are multiple
transactions with identical transaction names then they will all share
the same coordinates. This is deduplicated by preferring the transaction
with the lowest ID of the cluster for the center point.

* chore: Renaming blog post

* chore: Debugging things for post

* chore: Tweaks

* chore: Trying to wrap this up

* fix(blog): Fixing blog header text bug + progress on things

Added a script for generating the OG image for the blog post. This
script can eventually be adapted into other uses like generating images
automatically for all the pages on the docs site.

* chore: Tweak spacing on blog index

* feat(blog): Blog post progress!

* chore: Revisions up to tf-idf

* chore: Revisions up to DBSCAN

* chore: Revisions to end

* chore: Removing blog test

* chore: Tweaks

* chore: Fixing spelling and grammar mistakes

* chore: Fixing test